Refactored the code to be python3 compliant which meant changing the way some dictionaries worked, and other minor changes.
Added a whole bunch of TODO's as well to attempt to understand and further develop the code.

Added some comments and a couple doc strings, as well as reintroduced the 2**16 magic to the predictions.

Refactored the cad_driver.py file to read from csv's with pandas to avoid reading from stdin.